+++ projects/kavita-ng/i18n/src/public-api.ts
@@ -1,0 +1,3 @@
+export * from './langs';
+export * from './provide-localization';
+export * from './translator';